+++ packages/@ngtools/webpack/src/index.ts
@@ -1,0 +1,5 @@
+import 'reflect-metadata';
+
+export * from './plugin';
+export {ngcLoader as default} from './loader';
+export {PathsPlugin} from './paths-plugin';