+++ packages/angular-cli/blueprints/route/index.js
@@ -1,7 +1,0 @@
-module.exports = {
-  description: '',
-
-  install: function () {
-    throw 'Due to changes in the router, route generation has been temporarily disabled. You can find more information about the new router here: http://victorsavkin.com/post/145672529346/angular-router';
-  }
-};