+++ packages/angular-cli/commands/build.ts
@@ -11,6 +11,10 @@ export interface BuildOptions {
   supressSizes: boolean;
   baseHref?: string;
   aot?: boolean;
+  sourcemap?: boolean;
+  vendorChunk?: boolean;
+  verbose?: boolean;
+  progress?: boolean;
   i18nFile?: string;
   i18nFormat?: string;
   locale?: string;
@@ -35,8 +39,12 @@ const BuildCommand = Command.extend({
     { name: 'suppress-sizes', type: Boolean, default: false },
     { name: 'base-href',      type: String,  default: null, aliases: ['bh'] },
     { name: 'aot',            type: Boolean, default: false },
-    { name: 'i18n-file',       type: String, default: null },
-    { name: 'i18n-format',     type: String, default: null },
+    { name: 'sourcemap',      type: Boolean, default: true, aliases: ['sm'] },
+    { name: 'vendor-chunk',   type: Boolean, default: true },
+    { name: 'verbose',        type: Boolean, default: false },
+    { name: 'progress',       type: Boolean, default: true },
+    { name: 'i18n-file',      type: String, default: null },
+    { name: 'i18n-format',    type: String, default: null },
     { name: 'locale',         type: String, default: null }
   ],