+++ packages/angular-cli/ember-cli/lib/commands.js
@@ -1,0 +1,2 @@
+
+module.exports = { 'Unknown': require('./commands/unknown') };