+++ packages/angular-cli/ember-cli/lib/models/addon.js
@@ -19,11 +19,9 @@ var preprocessTemplates = p.preprocessTemplates;
 
 var AddonDiscovery  = require('../models/addon-discovery');
 var AddonsFactory   = require('../models/addons-factory');
-var CoreObject = require('core-object');
+var CoreObject = require('../ext/core-object');
 var Project = require('./project');
 
-var upstreamMergeTrees = require('broccoli-merge-trees');
-var Funnel     = require('broccoli-funnel');
 var walkSync   = require('walk-sync');
 
 
@@ -79,31 +77,6 @@ function Addon(parent, project) {
   }
   this.nodeModulesPath = nodeModulesPath(this.root);
 
-  this.treePaths = {
-    app:               'app',
-    styles:            'app/styles',
-    templates:         'app/templates',
-    addon:             'addon',
-    'addon-styles':    'addon/styles',
-    'addon-templates': 'addon/templates',
-    vendor:            'vendor',
-    'test-support':    'test-support',
-    'addon-test-support': 'addon-test-support',
-    public:            'public'
-  };
-
-  this.treeForMethods = {
-    app:               'treeForApp',
-    styles:            'treeForStyles',
-    templates:         'treeForTemplates',
-    'addon-templates': 'treeForAddonTemplates',
-    addon:             'treeForAddon',
-    vendor:            'treeForVendor',
-    'test-support':    'treeForTestSupport',
-    'addon-test-support': 'treeForAddonTestSupport',
-    public:            'treeForPublic'
-  };
-
   p.setupRegistry(this);
 
   if (!this.name) {
@@ -136,7 +109,6 @@ Addon.prototype._requireBuildPackages = function() {
   }
 
   this.transpileModules = deprecatedAddonFilters(this, 'this.transpileModules', 'broccoli-es6modules', function(tree, options) {
-    return new (require('broccoli-es6modules'))(tree, options);
   });
 
   this.pickFiles = deprecatedAddonFilters(this, 'this.pickFiles', 'broccoli-funnel', function(tree, options) {
@@ -148,7 +120,6 @@ Addon.prototype._requireBuildPackages = function() {
     return new Funnel(tree, options);
   });
 
-  this.mergeTrees = deprecatedAddonFilters(this, 'this.mergeTrees', 'broccoli-merge-trees', mergeTrees);
   this.walkSync = deprecatedAddonFilters(this, 'this.walkSync', 'node-walk-sync', walkSync);
 };
 
@@ -177,8 +148,6 @@ function deprecatedAddonFilters(addon, name, insteadUse, fn) {
   @return {tree} Modified tree
 */
 Addon.prototype.concatFiles = function(tree, options) {
-  options.sourceMapConfig = this.app.options.sourcemaps;
-  return require('broccoli-concat')(tree, options);
 };
 
 /**
@@ -278,33 +247,33 @@ Addon.prototype.treeFor = function treeFor(name) {
     trees.push(this.jshintAddonTree());
   }
 
-  return mergeTrees(trees.filter(Boolean), {
-    overwrite: true,
-    annotation: 'Addon#treeFor (' + this.name + ' - ' + name + ')'
-  });
+  // return mergeTrees(trees.filter(Boolean), {
+  //   overwrite: true,
+  //   annotation: 'Addon#treeFor (' + this.name + ' - ' + name + ')'
+  // });
 };
-
-/**
-  @private
-  @param {String} name
-  @method _treeFor
-  @return {tree}
-*/
-Addon.prototype._treeFor = function _treeFor(name) {
-  var treePath = path.resolve(this.root, this.treePaths[name]);
-  var treeForMethod = this.treeForMethods[name];
-  var tree;
-
-  // if (existsSync(treePath)) {
-  //   tree = this.treeGenerator(treePath);
-  // }
-
-  if (this[treeForMethod]) {
-    tree = this[treeForMethod](tree);
-  }
-
-  return tree;
-};
+//
+// /**
+//   @private
+//   @param {String} name
+//   @method _treeFor
+//   @return {tree}
+// */
+// Addon.prototype._treeFor = function _treeFor(name) {
+//   var treePath = path.resolve(this.root, this.treePaths[name]);
+//   var treeForMethod = this.treeForMethods[name];
+//   var tree;
+//
+//   // if (existsSync(treePath)) {
+//   //   tree = this.treeGenerator(treePath);
+//   // }
+//
+//   if (this[treeForMethod]) {
+//     tree = this[treeForMethod](tree);
+//   }
+//
+//   return tree;
+// };
 
 /**
   This method is called when the addon is included in a build. You