+++ packages/angular-cli/models/webpack-build-production.ts
@@ -1,6 +1,7 @@
 import * as path from 'path';
+import {CompressionPlugin} from '../lib/webpack/compression-plugin';
+
 const WebpackMd5Hash = require('webpack-md5-hash');
-const CompressionPlugin = require('compression-webpack-plugin');
 import * as webpack from 'webpack';
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 
@@ -14,14 +15,16 @@ declare module 'webpack' {
   }
 }
 
-export const getWebpackProdConfigPartial = function(projectRoot: string, appConfig: any) {
+export const getWebpackProdConfigPartial = function(projectRoot: string,
+                                                    appConfig: any,
+                                                    verbose: any) {
   const appRoot = path.resolve(projectRoot, appConfig.root);
   const styles = appConfig.styles
                ? appConfig.styles.map((style: string) => path.resolve(appRoot, style))
                : [];
   const cssLoaders = ['css-loader?sourcemap&minimize', 'postcss-loader'];
+
   return {
-    devtool: 'source-map',
     output: {
       path: path.resolve(projectRoot, appConfig.outDir),
       filename: '[name].[chunkhash].bundle.js',
@@ -58,19 +61,20 @@ export const getWebpackProdConfigPartial = function(projectRoot: string, appConf
       }),
       new webpack.optimize.UglifyJsPlugin(<any>{
         mangle: { screw_ie8 : true },
-        compress: { screw_ie8: true },
+        compress: { screw_ie8: true, warnings: verbose },
         sourceMap: true
       }),
       new CompressionPlugin({
           asset: '[path].gz[query]',
           algorithm: 'gzip',
-          test: /\.js$|\.html$/,
-          threshold: 10240,
-          minRatio: 0.8
+          test: /\.js$|\.html$|\.css$/,
+          threshold: 10240
       }),
       new webpack.LoaderOptionsPlugin({
+        test: /\.(css|scss|sass|less|styl)$/,
         options: {
           postcss: [
+            require('autoprefixer'),
             require('postcss-discard-comments')
           ]
         }